Error loading ftpservertest1.py
After downloading, the code does not run properly in downloads folder. To ensure it works, create a folder/directory in C:\ called "Shared". The whole path should be "C:\Shared". This will make sure the server works properly. I will create either a file that automatically does that, or debug this code.